+++ YouboraLib/transform/resourceparse/YBDashParser.m
@@ -8,7 +8,6 @@
 
 #import "YBDashParser.h"
 #import "YBRequest.h"
-#import "YBYouboraUtils.h"
 #import "YBDashFinalResourceParser.h"
 
 #define LOCATION @"Location"